+++ weblate/auth/data.py
@@ -72,6 +72,8 @@ PERMISSIONS = (
     # Translators: Permission name
     ("unit.review", gettext_noop("Review strings")),
     # Translators: Permission name
+    ("unit.bulk_edit", gettext_noop("Bulk edit strings")),
+    # Translators: Permission name
     ("unit.override", gettext_noop("Edit string when suggestions are enforced")),
     # Translators: Permission name
     ("unit.template", gettext_noop("Edit source strings")),
@@ -236,6 +238,10 @@ ROLES = (
         filter_perms("translation.", {"translation.auto"}),
     ),
     (
+        pgettext_noop("Access-control role", "Bulk editing"),
+        {"unit.bulk_edit"},
+    ),
+    (
         pgettext_noop("Access-control role", "Automatic translation"),
         {"translation.auto"},
     ),
@@ -305,6 +311,9 @@ ACL_GROUPS = {
     pgettext_noop(
         "Per-project access-control team name", "Screenshots"
     ): "Manage screenshots",
+    pgettext_noop(
+        "Per-project access-control team name", "Bulk editing"
+    ): "Bulk editing",
     pgettext_noop(
         "Per-project access-control team name", "Automatic translation"
     ): "Automatic translation",