+++ weblate/auth/utils.py
@@ -14,8 +14,10 @@ from django.db import IntegrityError
 from social_core.backends.utils import load_backends
 
 from weblate.auth.data import (
+    GLOBAL_PERM_NAMES,
     GLOBAL_PERMISSIONS,
     GROUPS,
+    PERMISSION_NAMES,
     PERMISSIONS,
     ROLES,
     SELECTION_ALL,
@@ -35,16 +37,15 @@ def get_auth_keys() -> set[str]:
 
 def is_django_permission(permission: str):
     """
-    Check whether permission looks like a Django one.
+    Check whether permission looks is a Django one.
 
-    Django permissions are <app>.<action>_<model>, while
-    Weblate ones are <scope>.<action> where action lacks underscores
-    with single exception of "add_more".
+    This is purely based on the list of permissions defined in Weblate.
     """
-    parts = permission.split(".", 1)
-    if len(parts) != 2:
-        return False
-    return "_" in parts[1] and parts[1] != "add_more"
+    return (
+        permission not in PERMISSION_NAMES
+        and permission not in GLOBAL_PERM_NAMES
+        and not permission.startswith(("meta:", "billing:"))
+    )
 
 
 def migrate_permissions_list(