+++ weblate/trans/tests/test_acl.py
@@ -347,21 +347,21 @@ class ACLTest(FixtureTestCase, RegistrationTestMixin):
         self.project.access_control = Project.ACCESS_PROTECTED
         self.project.translation_review = True
         self.project.save()
-        self.assertEqual(10 + billing_group, self.project.defined_groups.count())
+        self.assertEqual(11 + billing_group, self.project.defined_groups.count())
         self.project.access_control = Project.ACCESS_PRIVATE
         self.project.translation_review = True
         self.project.save()
-        self.assertEqual(10 + billing_group, self.project.defined_groups.count())
+        self.assertEqual(11 + billing_group, self.project.defined_groups.count())
         self.project.access_control = Project.ACCESS_CUSTOM
         self.project.save()
-        self.assertEqual(10 + billing_group, self.project.defined_groups.count())
+        self.assertEqual(11 + billing_group, self.project.defined_groups.count())
         self.project.access_control = Project.ACCESS_CUSTOM
         self.project.save()
-        self.assertEqual(10 + billing_group, self.project.defined_groups.count())
+        self.assertEqual(11 + billing_group, self.project.defined_groups.count())
         self.project.defined_groups.all().delete()
         self.project.access_control = Project.ACCESS_PRIVATE
         self.project.save()
-        self.assertEqual(10 + billing_group, self.project.defined_groups.count())
+        self.assertEqual(11 + billing_group, self.project.defined_groups.count())
         self.project.delete()
 
     def test_restricted_component(self) -> None: